requirements.txt
# 生产环境依赖
pyserial==3.5

# 开发环境依赖
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0 